move nanotime to common
